feat: 기존 axios 기반 fetching logic fetch Api 사용으로 migration
