+++ package.json
@@ -14,7 +14,6 @@
   },
   "dependencies": {
     "@types/styled-components": "^5.1.26",
-    "axios": "^1.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.11.1",
